requirements.txt
pandas
requests
yfinance
openai
numpy
bokeh
python-dotenv
